Add ProPub people
We need to tweak the credits block so those who worked on part 4 are properly recognized. This adds an optional note to the story metadata in the spreadsheet. If the note is present, it will be displayed and also trigger the display of all people with role of "PRO_PUBLICA".